move ModelVisitor to parsing from tidy
